Add a test case for the new ChildrensPlay play type.  We'll want to refactor this a bit to make these tests easier to add along with the new play types.
At the moment, adding a new type of Play involves subclassing Play and writing some functions, *as well as* instantiating a new play of that type, adding that play to a new invoice, and adding a new test that uses that invoice.  This is clunky as heck and can be improved on significantly.